feat(proposer): add proposer bin and match monorepo conventions
