+++ bin/proposer/README.md
@@ -1,0 +1,6 @@
+# `base-proposer-bin`
+
+TEE-based output proposer binary for OP Stack chains.
+
+Parses CLI arguments, builds a validated configuration, and delegates to
+`base_proposer::run()` for the full service lifecycle.
