+++ crates/proof/proposer/Cargo.toml
@@ -29,13 +29,13 @@ alloy-signer.workspace = true
 alloy-eips.workspace = true
 
 # Base Alloy
-base-alloy-rpc-types.workspace = true
 base-alloy-consensus = { workspace = true, features = ["std", "serde"] }
+base-alloy-rpc-types.workspace = true
 base-alloy-network.workspace = true
 
 # Enclave
-op-enclave-client.workspace = true
-op-enclave-core.workspace = true
+base-enclave-client.workspace = true
+base-enclave.workspace = true
 
 # Async
 async-trait.workspace = true
@@ -59,14 +59,20 @@ eyre.workspace = true
 base-cli-utils.workspace = true
 clap.workspace = true
 
-# Misc
-humantime.workspace = true
-jsonrpsee.workspace = true
+# Metrics
+metrics-exporter-prometheus.workspace = true
 metrics.workspace = true
+
+# TLS
+rustls = { workspace = true, features = ["ring"] }
+
+# Misc
+axum = { workspace = true, features = ["http1", "json", "tokio"] }
+moka = { workspace = true, features = ["future"] }
+jsonrpsee.workspace = true
+humantime.workspace = true
 backon.workspace = true
 bytes.workspace = true
-moka = { workspace = true, features = ["future"] }
-axum = { workspace = true, features = ["http1", "json", "tokio"] }
 hex.workspace = true
 url.workspace = true
