--- crates/proof/proposer/bin/proposer/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "proposer-bin"
-description = "Base Proposer Binary"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[[bin]]
-name = "proposer"
-path = "src/main.rs"
-
-[dependencies]
-metrics-exporter-prometheus.workspace = true
-tracing-subscriber.workspace = true
-alloy-primitives.workspace = true
-base-proposer.workspace = true
-rustls = { workspace = true, features = ["ring"] }
-tokio-util.workspace = true
-tracing.workspace = true
-metrics.workspace = true
-tokio.workspace = true
-clap.workspace = true
-eyre.workspace = true
