+++ crates/proof/proposer/src/contracts/aggregate_verifier.rs
@@ -67,7 +67,7 @@ pub trait AggregateVerifierClient: Send + Sync {
 }
 
 /// Concrete implementation backed by Alloy's sol-generated contract bindings.
-#[allow(missing_debug_implementations)]
+#[derive(Debug)]
 pub struct AggregateVerifierContractClient {
     provider: RootProvider,
 }
