+++ crates/proof/proposer/src/rpc/reth_client.rs
@@ -12,8 +12,8 @@ use alloy_provider::Provider;
 use alloy_rpc_types_eth::Header;
 use async_trait::async_trait;
 use backon::Retryable;
+use base_enclave::{AccountResult, executor::ExecutionWitness};
 use futures::stream::{self, StreamExt};
-use op_enclave_core::{AccountResult, executor::ExecutionWitness};
 
 use super::{
     error::{RpcError, RpcResult},
