+++ crates/proof/proposer/src/test_utils.rs
@@ -4,7 +4,7 @@ use std::sync::Arc;
 
 use alloy_primitives::{Address, B256, Bytes, U256};
 use async_trait::async_trait;
-use op_enclave_core::{
+use base_enclave::{
     AccountResult,
     executor::ExecutionWitness,
     types::config::{BlockId, Genesis, GenesisSystemConfig, PerChainConfig, RollupConfig},
